+++ src/generate/device.rs
@@ -1,4 +1,5 @@
 use crate::svd::{array::names, Device, Peripheral};
+use crate::Settings;
 use proc_macro2::{Span, TokenStream};
 use quote::{quote, ToTokens};
 
@@ -28,18 +29,24 @@ pub fn render(d: &Device, config: &Config, device_x: &mut String) -> Result<Toke
         }
     };
 
-    let settings = match config.settings.as_ref() {
+    let mut settings = match config.settings.as_ref() {
         #[cfg(feature = "yaml")]
         Some(settings) => {
             let file = std::fs::read_to_string(settings).context("could not read settings file")?;
-            Some(serde_yaml::from_str(&file).context("could not parse settings file")?)
+            serde_yaml::from_str(&file).context("could not parse settings file")?
         }
         #[cfg(not(feature = "yaml"))]
         Some(_) => {
             return Err(anyhow::anyhow!("Support for yaml config files is not available because svd2rust was compiled without the yaml feature"));
         }
-        None => None,
+        None => Settings::default(),
     };
+
+    // TODO: remove when `config.html_url` is removed
+    #[allow(deprecated)]
+    if let Some(html_url) = config.html_url.clone() {
+        settings.html_url = Some(html_url);
+    }
 
     // make_mod option explicitly disables inner attributes.
     if config.target == Target::Msp430 && !config.make_mod {
@@ -203,7 +210,7 @@ pub fn render(d: &Device, config: &Config, device_x: &mut String) -> Result<Toke
 
     match config.target {
         Target::RISCV => {
-            if settings.is_none() {
+            if settings.riscv_config.is_none() {
                 warn!("No settings file provided for RISC-V target. Using legacy interrupts rendering");
                 warn!("Please, consider migrating your PAC to riscv 0.12.0 or later");
                 out.extend(interrupt::render(
@@ -214,12 +221,7 @@ pub fn render(d: &Device, config: &Config, device_x: &mut String) -> Result<Toke
                 )?);
             } else {
                 debug!("Rendering RISC-V specific code");
-                out.extend(riscv::render(
-                    &d.peripherals,
-                    device_x,
-                    settings.as_ref().unwrap(),
-                    config,
-                )?);
+                out.extend(riscv::render(&d.peripherals, device_x, &settings, config)?);
             }
         }
         _ => {
@@ -241,16 +243,13 @@ pub fn render(d: &Device, config: &Config, device_x: &mut String) -> Result<Toke
             // Core peripherals are handled above
             continue;
         }
-        if config.target == Target::RISCV
-            && settings.is_some()
-            && riscv::is_riscv_peripheral(p, settings.as_ref().unwrap())
-        {
+        if config.target == Target::RISCV && riscv::is_riscv_peripheral(p, &settings) {
             // RISC-V specific peripherals are handled above
             continue;
         }
 
         debug!("Rendering peripheral {}", p.name);
-        let periph = peripheral::render(p, &index, config).with_context(|| {
+        let periph = peripheral::render(p, &index, &settings, config).with_context(|| {
             let group_name = p.group_name.as_deref().unwrap_or("No group name");
             let mut context_string =
                 format!("can't render peripheral '{}', group '{group_name}'", p.name);